# docs/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/profile": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UserController_listUsers",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete (deactivate) a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/roles": {
      "put": {
        "operationId": "UserController_setUserRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the user's assigned roles",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/locations": {
      "put": {
        "operationId": "UserController_setUserLocations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationAssignmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the user's assigned locations",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/change-password": {
      "post": {
        "operationId": "UserController_changePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password (user or allowed admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/reset-password": {
      "post": {
        "operationId": "UserController_adminResetPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin reset password",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Seed to Sale API",
    "description": "API documentation for Seed to Sale system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User UUID",
            "example": "11111111-1111-1111-1111-111111111111"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "name": {
            "type": "object",
            "description": "User full name",
            "example": "John Doe"
          },
          "roles": {
            "description": "List of role names assigned to the user",
            "example": [
              "licensee_admin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "locationIds": {
            "description": "List of location UUIDs the user is associated with",
            "example": [
              "33333333-3333-3333-3333-333333333333"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "Account status (active/inactive)",
            "example": "active"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp when the user was created"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp when the user was last updated"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Optional UUID for the user (useful for idempotent seeds)",
            "example": "11111111-1111-1111-1111-111111111111"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "description": "User password (write-only)",
            "example": "P@ssw0rd!",
            "minLength": 8,
            "writeOnly": true
          },
          "roleIds": {
            "description": "Array of role IDs assigned to the user (UUIDs).",
            "example": [
              "11111111-1111-1111-1111-111111111111"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "locationIds": {
            "description": "Array of location IDs the user is associated with",
            "example": [
              "33333333-3333-3333-3333-333333333333"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "parentLocationId": {
            "type": "string",
            "description": "Optional parent location id (UUID)",
            "example": "44444444-4444-4444-4444-444444444444"
          },
          "status": {
            "type": "string",
            "description": "Optional account status (e.g., \"active\", \"inactive\")",
            "example": "active"
          },
          "isActive": {
            "type": "boolean",
            "description": "Optional boolean toggling active state",
            "example": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "description": "New password. Provide only to change password via update (admin flow). Use change-password endpoint for user-initiated changes.",
            "minLength": 8,
            "writeOnly": true
          },
          "roleIds": {
            "description": "Array of role IDs to replace the user roles. To clear all roles pass an empty array.",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "locationIds": {
            "description": "Array of location IDs to replace the user locations. To clear pass an empty array.",
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "parentLocationId": {
            "type": "string",
            "description": "Optional parent location id (UUID)",
            "example": "44444444-4444-4444-4444-444444444444"
          },
          "status": {
            "type": "string",
            "description": "Optional account status",
            "example": "active"
          },
          "isActive": {
            "type": "boolean",
            "description": "Optional toggle to activate/deactivate account",
            "example": true
          }
        }
      },
      "RoleAssignmentDto": {
        "type": "object",
        "properties": {
          "roleIds": {
            "description": "Array of role UUIDs to set for the user (replaces existing roles).",
            "example": [
              "11111111-1111-1111-1111-111111111111"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "roleIds"
        ]
      },
      "LocationAssignmentDto": {
        "type": "object",
        "properties": {
          "locationIds": {
            "description": "Array of location UUIDs to set for the user (replaces existing locations).",
            "example": [
              "33333333-3333-3333-3333-333333333333"
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "locationIds"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password (required when a user changes their own password)",
            "example": "OldP@ssw0rd"
          },
          "newPassword": {
            "type": "string",
            "description": "New password (min length 8)",
            "example": "NewP@ssw0rd!"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "New password to set (admin-reset), minLength 8",
            "example": "NewP@ssw0rd!"
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
    }
  }
}